+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ New features and enhancements
 * Add annual global tas timeseries for CMIP6's model UKESM1-0-LL ssp585 r4i1p1f2 (:pull:`573`).
 * Add `strict_units` option to health checks. (:pull:`574`, :issue:`574`).
 * Add 10yr and 30yr frequencies support in CVs (:pull:`576`)
+* Add support for `omegaconf`'s interpolations in ``xscen.config.ConfigDict`` and the global dict `CONFIG`.
 
 v0.12.1 (2025-04-07)
 --------------------